LS: refactor
