+++ crates/cairo-lang-language-server/src/ide/hover/render/definition.rs
@@ -7,10 +7,10 @@ use cairo_lang_utils::Upcast;
 use tower_lsp::lsp_types::Hover;
 
 use crate::ide::hover::markdown_contents;
+use crate::ide::hover::render::markdown::{fenced_code_block, RULE};
 use crate::lang::db::AnalysisDatabase;
 use crate::lang::inspect::defs::{MemberDef, SymbolDef};
 use crate::lang::lsp::ToLsp;
-use crate::markdown::{fenced_code_block, RULE};
 
 /// Get declaration and documentation "definition" of an item referred by the given identifier.
 #[tracing::instrument(level = "trace", skip_all)]
