+++ crates/cairo-lang-language-server/src/ide/hover/render/mod.rs
@@ -3,3 +3,4 @@ pub use self::legacy::*;
 
 mod definition;
 mod legacy;
+mod markdown;
