+++ crates/cairo-lang-language-server/src/lsp/ext.rs
@@ -1,8 +1,28 @@
 //! CairoLS extensions to the Language Server Protocol.
 
+use serde::{Deserialize, Serialize};
 use tower_lsp::lsp_types::notification::Notification;
 use tower_lsp::lsp_types::request::Request;
-use tower_lsp::lsp_types::TextDocumentPositionParams;
+use tower_lsp::lsp_types::{TextDocumentPositionParams, Url};
+
+/// Provides content of virtual file from the database.
+pub struct ProvideVirtualFile;
+
+#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
+pub struct ProvideVirtualFileRequest {
+    pub uri: Url,
+}
+
+#[derive(Debug, Eq, PartialEq, Clone, Deserialize, Serialize)]
+pub struct ProvideVirtualFileResponse {
+    pub content: Option<String>,
+}
+
+impl Request for ProvideVirtualFile {
+    type Params = ProvideVirtualFileRequest;
+    type Result = ProvideVirtualFileResponse;
+    const METHOD: &'static str = "vfs/provide";
+}
 
 /// Collects information about all Cairo crates that are currently being analyzed.
 pub struct ViewAnalyzedCrates;
@@ -13,7 +33,7 @@ impl Request for ViewAnalyzedCrates {
     const METHOD: &'static str = "cairo/viewAnalyzedCrates";
 }
 
-/// Provides string with code after macros expansion
+/// Provides string with code after macros expansion.
 pub struct ExpandMacro;
 
 impl Request for ExpandMacro {
