+++ crates/cairo-lang-language-server/src/lsp/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod capabilities;
 pub mod ext;
+mod methods;
