+++ crates/cairo-lang-language-server/src/server/mod.rs
@@ -1,1 +1,2 @@
+pub mod commands;
 pub mod notifier;
